🐛 Fix LimineFramebuffer struct layout (BitsPerPixel was ulong, should be ushort)
BitsPerPixel is uint16_t in the Limine spec but was declared as ulong (8 bytes)
instead of ushort (2 bytes). This shifted EdidSize and Edid to wrong offsets,
causing a page fault on x64 when reading EDID data.

+++ src/Cosmos.Kernel.Boot.Limine/LimineFramebuffer.cs
@@ -61,7 +61,7 @@ public readonly unsafe struct LimineFramebuffer
     public readonly ulong Width;
     public readonly ulong Height;
     public readonly ulong Pitch;
-    public readonly ulong BitsPerPixel;
+    public readonly ushort BitsPerPixel;
     public readonly LimineFbMemoryModel MemoryModel;
     public readonly byte RedMaskSize;
     public readonly byte RedMaskShift;
